fix: emission TX creation, Gulf Stream forwarding, explorer sorting and local badges
CRITICAL FIXES:
- Emission: Use load_microblock_auto_format() to correctly load transactions from EfficientMicroBlock
- Emission: Implement delayed reward (current_epoch - 2) to avoid race condition with MacroBlock creation
- Gulf Stream: Deterministic producer calculation using select_microblock_producer() for next blocks
- Gulf Stream: Handle rotation boundaries by forwarding to both current and next producers

EXPLORER IMPROVEMENTS:
- Sort transactions by block height (primary) instead of timestamp
- Add Heartbeat and Validator transaction type badges
- Multi-select type filter with Clear button
- Fix gas_price column: BIGINT -> NUMERIC(20,0) to support u64::MAX

DUPLICATE TX FIX:
- Skip duplicate NodeActivation TX for genesis nodes (already registered in block 0)
- Add QNET-BOOT- prefix check in quantum_crypto.rs

UI/UX:
- Download App Store and Google Play badges locally (no external CDN dependency)
- Adjust badge sizes for visual consistency

+++ applications/qnet-explorer/frontend/migrations/001_init.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS transactions (
     nonce NUMERIC(20,0) NOT NULL,
     block BIGINT NOT NULL,
     timestamp BIGINT NOT NULL,
-    gas_price BIGINT NOT NULL DEFAULT 0,
+    gas_price NUMERIC(20,0) NOT NULL DEFAULT 0,
     gas_limit BIGINT NOT NULL DEFAULT 0,
     signature TEXT,
     public_key TEXT,
